+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.7.0
+
+- Changed: preserve official output format
+- Changed: customizable token estimation
+- Changed: default api version at `2024-02-01`
+
 # v0.6.1
 
 - Changed: default api version at `2024-02-15-preview`